move VMTest into vmtest_helper.py
